+++ kyuubi-main/src/test/scala/org/apache/kyuubi/operation/KyuubiOperationPerUserSuite.scala
@@ -19,7 +19,7 @@ package org.apache.kyuubi.operation
 
 import org.apache.kyuubi.config.KyuubiConf
 
-class KyuubiOperationPerUserSuite extends SharedKyuubiOperationContext with JDBCTests {
+class KyuubiOperationPerUserSuite extends JDBCTests with SharedKyuubiOperationContext {
 
   override protected def jdbcUrl: String = getJdbcUrl
